requirements.txt
bs4==0.0.1
mysql-connector-python==8.0.29
requests==2.27.1